refactor(tests): consolidate test suite, remove duplicates
Test consolidation (Phase 6):
- Merge CLI utilities: test_cli_formatting.py + test_cli_helpers.py +
  test_cli_container.py → test_cli_utilities.py (75 tests)
- Remove 9 duplicate test classes from test_db_store.py
- Delete superseded files: test_editor.py (→test_cli_editor.py),
  test_drive_config.py (trivial), test_retries.py (→test_drive_resilience.py),
  test_qdrant.py (→test_qdrant_resilience.py + test_search_index.py)

Result: 102 → 96 test files, 2149 → 2125 tests (no lost coverage)
