+++ tests/test_cli_utilities.py
@@ -1,26 +1,44 @@
-"""Tests for polylogue.cli.helpers module.
+"""Consolidated tests for CLI utility modules.
 
-Coverage targets:
-- fail: raises SystemExit with message
-- is_declarative: checks POLYLOGUE_DECLARATIVE env var
-- source_state_path: returns correct XDG path
-- load_last_source/save_last_source: persistence of source selection
-- maybe_prompt_sources: interactive source selection
-- load_effective_config: config loading
-- resolve_sources: source name validation
-- print_summary: output formatting
-- latest_render_path: finds most recent render file
+This file consolidates tests from:
+- test_cli_formatting.py (25 tests)
+- test_cli_helpers.py (48 tests)
+- test_cli_container.py (2 tests)
+
+Coverage includes:
+- CLI formatting functions (format_counts, format_cursors, format_source_label, etc.)
+- CLI helper functions (fail, resolve_sources, load_effective_config, etc.)
+- Storage repository factory functions
 """
 
 from __future__ import annotations
 
 import json
+import sys
+from io import StringIO
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
 
 from polylogue.cli import helpers
+from polylogue.cli.formatting import (
+    announce_plain_mode,
+    format_counts,
+    format_cursors,
+    format_index_status,
+    format_source_label,
+    format_sources_summary,
+    should_use_plain,
+)
+from polylogue.config import Source
+from polylogue.storage.backends.sqlite import create_default_backend
+from polylogue.storage.repository import StorageRepository
+
+
+# ============================================================================
+# Fixtures
+# ============================================================================
 
 
 @pytest.fixture
@@ -49,6 +67,239 @@ def helpers_workspace(tmp_path, monkeypatch):
         "state_dir": state_dir,
         "config_dir": config_dir,
     }
+
+
+# ============================================================================
+# Formatting Tests
+# ============================================================================
+
+
+class TestShouldUsePlain:
+    """Test should_use_plain function."""
+
+    def test_explicit_plain_true(self):
+        """Explicit plain=True returns True."""
+        assert should_use_plain(plain=True, interactive=False) is True
+
+    def test_explicit_interactive_true(self):
+        """Explicit interactive=True returns False."""
+        assert should_use_plain(plain=False, interactive=True) is False
+
+    def test_plain_takes_precedence_over_interactive(self):
+        """plain=True takes precedence over interactive=True."""
+        assert should_use_plain(plain=True, interactive=True) is True
+
+    def test_env_var_force_plain(self, monkeypatch):
+        """POLYLOGUE_FORCE_PLAIN env var enables plain mode."""
+        monkeypatch.setenv("POLYLOGUE_FORCE_PLAIN", "1")
+        assert should_use_plain(plain=False, interactive=False) is True
+
+    def test_env_var_false_values(self, monkeypatch):
+        """POLYLOGUE_FORCE_PLAIN with 0/false/no doesn't force plain."""
+        for val in ("0", "false", "no", "False", "NO"):
+            monkeypatch.setenv("POLYLOGUE_FORCE_PLAIN", val)
+            # Result depends on TTY status, but env var doesn't force plain
+            # We can't easily test TTY, so we just verify no crash
+            result = should_use_plain(plain=False, interactive=False)
+            assert isinstance(result, bool)
+
+    def test_non_tty_returns_true(self, monkeypatch):
+        """Non-TTY environment returns True (plain mode)."""
+        monkeypatch.delenv("POLYLOGUE_FORCE_PLAIN", raising=False)
+        # Mock stdout/stderr as non-TTY
+        with patch.object(sys.stdout, "isatty", return_value=False):
+            with patch.object(sys.stderr, "isatty", return_value=False):
+                assert should_use_plain(plain=False, interactive=False) is True
+
+
+class TestAnnouncePlainMode:
+    """Test announce_plain_mode function."""
+
+    def test_writes_to_stderr(self):
+        """Writes announcement to stderr."""
+        captured = StringIO()
+        with patch.object(sys, "stderr", captured):
+            announce_plain_mode()
+        output = captured.getvalue()
+        assert "Plain output active" in output
+        assert "--interactive" in output
+
+
+class TestFormatCursors:
+    """Test format_cursors function."""
+
+    def test_empty_cursors_returns_none(self):
+        """Empty cursors dict returns None."""
+        assert format_cursors({}) is None
+
+    def test_file_count_displayed(self):
+        """File count is displayed."""
+        result = format_cursors({"inbox": {"file_count": 10}})
+        assert result is not None
+        assert "10 files" in result
+        assert "inbox" in result
+
+    def test_error_count_highlighted(self):
+        """Error count is displayed when non-zero."""
+        result = format_cursors({"source": {"file_count": 5, "error_count": 2}})
+        assert result is not None
+        assert "2 errors" in result
+
+    def test_zero_error_count_not_shown(self):
+        """Zero error count is not displayed."""
+        result = format_cursors({"source": {"file_count": 5, "error_count": 0}})
+        assert result is not None
+        assert "errors" not in result
+
+    def test_latest_mtime_formatted(self):
+        """Latest mtime is formatted as timestamp."""
+        result = format_cursors({"source": {"latest_mtime": 1704067200}})
+        assert result is not None
+        assert "latest" in result
+        # Should contain ISO-ish format
+        assert "202" in result  # Year prefix
+
+    def test_latest_file_name_shown(self):
+        """Latest file name is shown when mtime not available."""
+        result = format_cursors({"source": {"latest_file_name": "chat.json"}})
+        assert result is not None
+        assert "latest chat.json" in result
+
+    def test_latest_path_fallback(self):
+        """Path basename used as fallback for latest label."""
+        result = format_cursors({"source": {"latest_path": "/some/dir/export.json"}})
+        assert result is not None
+        assert "latest export.json" in result
+
+    def test_multiple_cursors(self):
+        """Multiple cursors are joined with semicolons."""
+        result = format_cursors({
+            "inbox": {"file_count": 5},
+            "drive": {"file_count": 3},
+        })
+        assert result is not None
+        assert "inbox" in result
+        assert "drive" in result
+        assert ";" in result
+
+
+class TestFormatCounts:
+    """Test format_counts function."""
+
+    def test_conversations_and_messages(self):
+        """Shows conversations and messages count."""
+        result = format_counts({"conversations": 10, "messages": 100})
+        assert "10 conv" in result
+        assert "100 msg" in result
+
+    def test_rendered_shown_when_nonzero(self):
+        """Rendered count shown when non-zero."""
+        result = format_counts({"conversations": 5, "messages": 50, "rendered": 5})
+        assert "5 rendered" in result
+
+    def test_rendered_not_shown_when_zero(self):
+        """Rendered count not shown when zero."""
+        result = format_counts({"conversations": 5, "messages": 50, "rendered": 0})
+        assert "rendered" not in result
+
+    def test_missing_keys_default_to_zero(self):
+        """Missing keys default to zero."""
+        result = format_counts({})
+        assert "0 conv" in result
+        assert "0 msg" in result
+
+
+class TestFormatIndexStatus:
+    """Test format_index_status function."""
+
+    def test_ingest_stage_skipped(self):
+        """Ingest stage shows skipped."""
+        assert format_index_status("ingest", True, None) == "Index: skipped"
+
+    def test_render_stage_skipped(self):
+        """Render stage shows skipped."""
+        assert format_index_status("render", False, None) == "Index: skipped"
+
+    def test_index_error(self):
+        """Index error is reported."""
+        assert format_index_status("full", True, "connection failed") == "Index: error"
+
+    def test_indexed_ok(self):
+        """Indexed flag True shows ok."""
+        assert format_index_status("full", True, None) == "Index: ok"
+
+    def test_not_indexed_up_to_date(self):
+        """Not indexed shows up-to-date."""
+        assert format_index_status("full", False, None) == "Index: up-to-date"
+
+
+class TestFormatSourceLabel:
+    """Test format_source_label function."""
+
+    def test_source_differs_from_provider(self):
+        """Shows source/provider when they differ."""
+        result = format_source_label("inbox", "claude")
+        assert result == "inbox/claude"
+
+    def test_source_same_as_provider(self):
+        """Shows just source when same as provider."""
+        result = format_source_label("claude", "claude")
+        assert result == "claude"
+
+    def test_none_source(self):
+        """None source shows provider name."""
+        result = format_source_label(None, "chatgpt")
+        assert result == "chatgpt"
+
+
+class TestFormatSourcesSummary:
+    """Test format_sources_summary function."""
+
+    def test_empty_sources(self):
+        """Empty list returns 'none'."""
+        assert format_sources_summary([]) == "none"
+
+    def test_path_source(self):
+        """Source with path shows name."""
+        source = Source(name="inbox", path=Path("/inbox"))
+        result = format_sources_summary([source])
+        assert "inbox" in result
+        assert "(drive)" not in result
+
+    def test_drive_source(self):
+        """Source with folder shows (drive) tag."""
+        source = Source(name="gemini", folder="folder-id")
+        result = format_sources_summary([source])
+        assert "gemini (drive)" in result
+
+    def test_missing_source(self):
+        """Source without path or folder shows (missing)."""
+        # Note: Source validation prevents creating such objects normally
+        # This tests defensive code handling edge cases via mock
+        from unittest.mock import MagicMock
+
+        source = MagicMock()
+        source.name = "broken"
+        source.path = None
+        source.folder = None
+        result = format_sources_summary([source])
+        assert "broken (missing)" in result
+
+    def test_truncates_long_lists(self):
+        """Lists > 8 items are truncated."""
+        sources = [
+            Source(name=f"source{i}", path=Path(f"/src{i}"))
+            for i in range(12)
+        ]
+        result = format_sources_summary(sources)
+        assert "+4 more" in result
+        # Should have 8 names plus the "+4 more"
+        assert result.count(",") == 8  # 9 items = 8 commas
+
+
+# ============================================================================
+# Helper Tests
+# ============================================================================
 
 
 class TestFail:
@@ -493,3 +744,28 @@ class TestLoadEffectiveConfig:
         config = helpers.load_effective_config(env)
         assert config is not None
         assert hasattr(config, "sources")
+
+
+# ============================================================================
+# Container Tests
+# ============================================================================
+
+
+def test_create_storage_repository() -> None:
+    """Test creating storage repository returns StorageRepository instance."""
+    backend = create_default_backend()
+    repository = StorageRepository(backend=backend)
+
+    assert isinstance(repository, StorageRepository)
+    assert hasattr(repository, "_write_lock")
+
+
+def test_create_storage_repository_independent_instances() -> None:
+    """Test that each call creates a new independent repository instance."""
+    backend1 = create_default_backend()
+    backend2 = create_default_backend()
+    repo1 = StorageRepository(backend=backend1)
+    repo2 = StorageRepository(backend=backend2)
+
+    assert repo1 is not repo2
+    assert repo1._write_lock is not repo2._write_lock
