+++ tests/test_db_store.py
@@ -64,167 +64,6 @@ class TestConnectionCommitAndRollback:
                 assert row["conversation_id"] == "c1"
 
 
-class TestAttachmentRefCounting:
-    """Test attachment ref counting and cleanup."""
-
-    def test_attachment_ref_increment_on_store(self, tmp_path):
-        """Storing conversation with attachments increments ref_count."""
-        db_path = tmp_path / "test.db"
-
-        conv = make_conversation("c1")
-        msg = make_message("m1", "c1")
-        att = make_attachment("att1", "c1", "m1", mime_type="image/png")
-
-        with open_connection(db_path) as conn:
-            counts = store_records(conversation=conv, messages=[msg], attachments=[att], conn=conn)
-
-        assert counts["attachments"] == 1
-
-        # Verify ref_count
-        with open_connection(db_path) as conn:
-            row = conn.execute("SELECT ref_count FROM attachments WHERE attachment_id = ?", ("att1",)).fetchone()
-            assert row["ref_count"] == 1
-
-    def test_same_attachment_referenced_twice(self, tmp_path):
-        """Same attachment ID referenced by two messages has ref_count=2."""
-        db_path = tmp_path / "test.db"
-
-        conv = make_conversation("c1")
-        msg1 = make_message("m1", "c1", text="Message 1")
-        msg2 = make_message("m2", "c1", role="assistant", text="Message 2")
-
-        # Same attachment in both messages
-        att1 = make_attachment("att1", "c1", "m1", mime_type="image/png")
-        att2 = make_attachment("att1", "c1", "m2", mime_type="image/png")
-
-        with open_connection(db_path) as conn:
-            counts = store_records(
-                conversation=conv,
-                messages=[msg1, msg2],
-                attachments=[att1, att2],
-                conn=conn,
-            )
-
-        assert counts["attachments"] == 2
-
-        # Verify ref_count is 2
-        with open_connection(db_path) as conn:
-            row = conn.execute("SELECT ref_count FROM attachments WHERE attachment_id = ?", ("att1",)).fetchone()
-            assert row["ref_count"] == 2
-
-            # Verify two refs in table
-            ref_count = conn.execute(
-                "SELECT COUNT(*) FROM attachment_refs WHERE attachment_id = ?", ("att1",)
-            ).fetchone()[0]
-            assert ref_count == 2
-
-    def test_attachment_ref_decrement_on_update(self, tmp_path):
-        """Updating conversation to remove attachment decrements ref."""
-        db_path = tmp_path / "test.db"
-
-        # First version: conversation with attachment
-        conv1 = make_conversation("c1", title="Version 1", content_hash="hash1")
-        msg1 = make_message("m1", "c1", text="Message 1", timestamp="2024-01-01T00:00:00Z")
-        att1 = make_attachment("att1", "c1", "m1", mime_type="image/png")
-
-        with open_connection(db_path) as conn:
-            store_records(conversation=conv1, messages=[msg1], attachments=[att1], conn=conn)
-
-        # Verify attachment exists with ref_count=1
-        with open_connection(db_path) as conn:
-            row = conn.execute("SELECT ref_count FROM attachments WHERE attachment_id = ?", ("att1",)).fetchone()
-            assert row["ref_count"] == 1
-
-        # Second version: same conversation, no attachments
-        conv2 = make_conversation("c1", title="Version 2", content_hash="hash2")
-
-        with open_connection(db_path) as conn:
-            store_records(conversation=conv2, messages=[msg1], attachments=[], conn=conn)
-
-        # Verify attachment was deleted (ref_count dropped to 0)
-        with open_connection(db_path) as conn:
-            row = conn.execute("SELECT * FROM attachments WHERE attachment_id = ?", ("att1",)).fetchone()
-            assert row is None
-
-
-class TestStoreConversationUpsert:
-    """Test conversation upsert logic."""
-
-    def test_store_records_inserts_new_conversation(self, tmp_path):
-        """store_records() inserts new conversation."""
-        db_path = tmp_path / "test.db"
-
-        conv = make_conversation("c1", title="New Conversation")
-        msg = make_message("m1", "c1", text="Hello")
-
-        with open_connection(db_path) as conn:
-            counts = store_records(conversation=conv, messages=[msg], attachments=[], conn=conn)
-
-        assert counts["conversations"] == 1
-        assert counts["messages"] == 1
-
-        # Verify in database
-        with open_connection(db_path) as conn:
-            row = conn.execute("SELECT * FROM conversations WHERE conversation_id = ?", ("c1",)).fetchone()
-            assert row is not None
-            assert row["title"] == "New Conversation"
-
-    def test_store_records_skips_duplicate_on_same_hash(self, tmp_path):
-        """store_records() skips conversation with same content_hash."""
-        db_path = tmp_path / "test.db"
-
-        conv = make_conversation("c1", title="Same Title", content_hash="samehash")
-
-        with open_connection(db_path) as conn:
-            # First insert
-            counts1 = store_records(conversation=conv, messages=[], attachments=[], conn=conn)
-            assert counts1["conversations"] == 1
-
-            # Identical second insert
-            counts2 = store_records(conversation=conv, messages=[], attachments=[], conn=conn)
-            assert counts2["conversations"] == 0
-            assert counts2["skipped_conversations"] == 1
-
-    def test_store_records_updates_on_different_hash(self, tmp_path):
-        """store_records() updates conversation when content_hash differs."""
-        db_path = tmp_path / "test.db"
-
-        # First version
-        conv1 = make_conversation("c1", title="Original Title", content_hash="hash1")
-
-        with open_connection(db_path) as conn:
-            store_records(conversation=conv1, messages=[], attachments=[], conn=conn)
-
-        # Second version with different content
-        conv2 = make_conversation("c1", title="Updated Title", content_hash="hash2")
-
-        with open_connection(db_path) as conn:
-            counts = store_records(conversation=conv2, messages=[], attachments=[], conn=conn)
-
-        assert counts["conversations"] == 1
-
-        # Verify update
-        with open_connection(db_path) as conn:
-            row = conn.execute("SELECT * FROM conversations WHERE conversation_id = ?", ("c1",)).fetchone()
-            assert row["title"] == "Updated Title"
-            assert row["content_hash"] == "hash2"
-
-    def test_store_records_multiple_conversations(self, tmp_path):
-        """store_records() handles multiple conversations independently."""
-        db_path = tmp_path / "test.db"
-
-        conversations = [make_conversation(f"c{i}", title=f"Conversation {i}", content_hash=f"hash{i}") for i in range(5)]
-
-        with open_connection(db_path) as conn:
-            for conv in conversations:
-                store_records(conversation=conv, messages=[], attachments=[], conn=conn)
-
-        # Verify all inserted
-        with open_connection(db_path) as conn:
-            count = conn.execute("SELECT COUNT(*) FROM conversations").fetchone()[0]
-            assert count == 5
-
-
 class TestThreadSafety:
     """Test thread-local connection safety."""
 
@@ -299,73 +138,6 @@ class TestThreadSafety:
 
             msg_count = conn.execute("SELECT COUNT(*) FROM messages").fetchone()[0]
             assert msg_count == 60  # 20 * 3
-
-
-class TestAttachmentPruning:
-    """Test attachment pruning and cleanup logic."""
-
-    def test_prune_removes_orphaned_attachments(self, tmp_path):
-        """Pruning with empty keep_set removes all attachments."""
-        db_path = tmp_path / "test.db"
-
-        conv = make_conversation("c1")
-        msg = make_message("m1", "c1")
-        att = make_attachment("att1", "c1", "m1", mime_type="image/png")
-
-        with open_connection(db_path) as conn:
-            store_records(conversation=conv, messages=[msg], attachments=[att], conn=conn)
-
-            # Verify attachment exists
-            row = conn.execute("SELECT * FROM attachments WHERE attachment_id = ?", ("att1",)).fetchone()
-            assert row is not None
-
-            # Prune with empty keep set
-            _prune_attachment_refs(conn, "c1", set())
-
-            # Verify attachment deleted
-            row = conn.execute("SELECT * FROM attachments WHERE attachment_id = ?", ("att1",)).fetchone()
-            assert row is None
-
-    def test_prune_keeps_specified_refs(self, tmp_path):
-        """Pruning with keep_set preserves specified refs."""
-        db_path = tmp_path / "test.db"
-
-        conv = make_conversation("c1")
-        msg1 = make_message("m1", "c1", text="Message 1")
-        msg2 = make_message("m2", "c1", role="assistant", text="Message 2")
-        att1 = make_attachment("att1", "c1", "m1", mime_type="image/png")
-        att2 = make_attachment("att2", "c1", "m2", mime_type="image/jpeg", size_bytes=2048)
-
-        with open_connection(db_path) as conn:
-            store_records(
-                conversation=conv,
-                messages=[msg1, msg2],
-                attachments=[att1, att2],
-                conn=conn,
-            )
-
-            ref_id1 = _make_ref_id("att1", "c1", "m1")
-            ref_id2 = _make_ref_id("att2", "c1", "m2")
-
-            # Verify both exist
-            count_before = conn.execute(
-                "SELECT COUNT(*) FROM attachments"
-            ).fetchone()[0]
-            assert count_before == 2
-
-            # Prune, keeping only att1
-            _prune_attachment_refs(conn, "c1", {ref_id1})
-
-            # Verify att1 kept, att2 deleted
-            count_after = conn.execute(
-                "SELECT COUNT(*) FROM attachments"
-            ).fetchone()[0]
-            assert count_after == 1
-
-            att1_row = conn.execute("SELECT * FROM attachments WHERE attachment_id = ?", ("att1",)).fetchone()
-            att2_row = conn.execute("SELECT * FROM attachments WHERE attachment_id = ?", ("att2",)).fetchone()
-            assert att1_row is not None
-            assert att2_row is None
 
 
 class TestSchemaAndMigration:
